+++ system/ui/widgets/scroller.py
@@ -1,6 +1,6 @@
 import pyray as rl
+from openpilot.system.ui.lib.scroll_panel import GuiScrollPanel
 from openpilot.system.ui.widgets import Widget
-from openpilot.system.ui.widgets.scroll_panel import GuiScrollPanel
 
 ITEM_SPACING = 40
 LINE_COLOR = rl.GRAY
